Update tests to reflect new focus management style
