+++ js/__mocks__/mediaMock.js
@@ -1,0 +1,9 @@
+const media = {
+  addEventListener: jest.fn(),
+  setVolume: jest.fn(),
+  setBalance: jest.fn(),
+  setPreamp: jest.fn(),
+  getAnalyser: () => null,
+  on: jest.fn(),
+};
+export default media;
